src/requirements.txt
android-backup==0.2.0
appdirs==1.4.4
attrs==19.3.0
cffi==1.14.0
click==7.1.2
construct==2.10.56
croniter==0.3.34
cryptography==2.9.2
ifaddr==0.1.7
importlib-metadata==1.7.0
natsort==7.0.1
netifaces==0.10.9
prometheus-client==0.8.0
pycparser==2.20
python-dateutil==2.8.1
python-miio==0.5.2.1
pytz==2019.3
six==1.15.0
tqdm==4.47.0
zeroconf==0.25.1
zipp==3.1.0
